# Projects/Project-3/backend/docs/openapi/files.yaml
openapi: 3.0.0
paths:
  /files:
    get:
      operationId: FilesController_list
      parameters:
        - name: cursor
          required: false
          in: query
          description: Opaque; take it from `meta.nextCursor` of the previous page.
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: sort
          required: false
          in: query
          description: A leading `-` is descending.
          schema:
            default: -createdAt
            type: string
            enum:
              - -createdAt
              - createdAt
        - name: mimeType
          required: false
          in: query
          schema:
            type: string
            enum:
              - text/csv
              - application/vnd.ms-excel
              - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
        - name: visibility
          required: false
          in: query
          schema:
            type: string
            enum:
              - company
              - restricted
        - name: uploaderId
          required: false
          in: query
          schema:
            format: uuid
            type: string
        - name: uploadedAfter
          required: false
          in: query
          description: Uploaded at or after this instant.
          schema:
            format: date-time
            type: string
        - name: uploadedBefore
          required: false
          in: query
          description: Uploaded before this instant.
          schema:
            format: date-time
            type: string
        - name: allVersions
          required: false
          in: query
          description: >-
            By default each file is listed once, as its newest version. `true`
            lists every version you can see.
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilePageDto'
      security: &ref_0
        - bearer: []
      tags: &ref_1
        - files
      summary: List files
      description: >-
        The files the caller may see: everything company-wide, plus restricted
        files they uploaded or were granted, and every file for an admin.
        Cursor-paginated, newest first (`sort=createdAt` reverses it); follow
        `meta.nextCursor` until `hasMore` is false. Filters (`mimeType`,
        `visibility`, `uploaderId`, `uploadedAfter`, `uploadedBefore`) only ever
        narrow that set. Rows never include where the bytes are stored or who a
        restricted file is shared with.


        A file that has several **versions** is listed once, as its newest
        version (`isLatest`); `allVersions=true` lists every version you can
        see. `datasetId` is shared by all the versions of one file.
    post:
      operationId: FilesController_upload
      parameters:
        - name: Idempotency-Key
          in: header
          description: >-
            A UUID. A retry with the same key and the same file replays the
            first response instead of uploading twice.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadFileBodyDoc'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
      security: *ref_0
      tags: *ref_1
      summary: Upload a spreadsheet
      description: >-
        Multipart upload of a CSV, XLS or XLSX file up to 25 MB, with a
        `visibility` of `company` (the default) or `restricted` and, for a
        restricted file, `grantedUserIds` (repeat the field or send a JSON
        array; they must be active members of the company — get their ids from
        `GET /companies/me/members`). The type is decided from the file's
        **bytes**, never from its name or `Content-Type`: an executable renamed
        `report.csv` is refused with 400, as is a Word document renamed `.xls`.
        A refused, invalid or failed upload stores nothing and consumes no
        quota.


        **Quota.** Free (10 files per billing period) and Basic (100) answer
        **402** once the period's quota is used, naming the plan, the count, the
        cap and the date the quota resets — the upload never starts. Premium
        (1000 included) always accepts; a file past the 1000th carries an
        `X-Gridline-Quota-Warning` header and a $0.50 overage charge. Deleting a
        file does not refund its quota.


        **Idempotency.** Send an `Idempotency-Key` (a UUID) and a retry with the
        same key and the same file replays the first response — marked
        `Idempotent-Replayed: true`, warning header included — instead of
        uploading and counting twice. The same key with a different request is
        422; while the first request is still running, 409. A request that
        failed is not remembered.


        Each upload queues a data-quality report and is recorded in the audit
        log. Returns 413 above 25 MB.
  /files/{id}/versions:
    get:
      operationId: FilesController_versions
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileVersionPageDto'
      security: *ref_0
      tags: *ref_1
      summary: List a file's versions
      description: >-
        Every version of the file `id` belongs to that the caller may see,
        newest first, offset-paginated. Works from any of its versions. Versions
        the caller cannot see are left out, and a file you cannot see at all is
        a 404.
    post:
      operationId: FilesController_uploadVersion
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          description: >-
            A UUID. A retry with the same key and the same file replays the
            first response instead of uploading twice.
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadVersionBodyDoc'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
      security: *ref_0
      tags: *ref_1
      summary: Upload a new version of a file
      description: >-
        Adds the next version to the file `id` belongs to (any of its versions
        will do): version 2, then 3, and so on. Multipart, with the same `file`
        part and the same rules as `POST /files` — the type is read from the
        bytes, up to 25 MB, `Idempotency-Key` is honoured — and it is a real
        upload: **it counts toward the period's file quota** and the plan's
        limit applies (402 past it, exactly like a new file), and it gets its
        own data-quality report.


        The new version becomes the file's **latest**. It inherits who can see
        the file — the visibility and the grants of the previous latest version,
        plus that version's uploader, so nobody loses the file because someone
        else uploaded the next version — and takes **no fields** of its own
        (sending one is a 400). After that each version's access is its own:
        `PATCH /files/{id}` changes one version only.


        Only the uploader or an admin may add a version: someone who can see the
        file but did not upload it gets 403, and a file you cannot see (or that
        was deleted) is a 404. Each plan keeps a number of versions per file
        (Free 5, Basic 50, Premium unlimited); beyond it the answer is 409
        naming the plan and the number, and deleting a version makes room.
        Version numbers are never reused. Needs the `files:write` scope for an
        API key.


        When the new version drops or retypes a column that the previous one
        had, the uploader and the admins get a `dataset.schema_changed`
        notification once its report is built.
  /files/{id}/compare/{otherId}:
    get:
      operationId: FilesController_compare
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: otherId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonDto'
      security: *ref_0
      tags: *ref_1
      summary: Compare two versions of a file
      description: >-
        What changed from `id` to `otherId` (normally an older version to a
        newer one), worked out from their stored data-quality reports — nothing
        is read from the files again. Both must be visible to you (a file you
        cannot see is a 404) and be versions of the same file (422 otherwise),
        and both reports must be `ready` (409 while one is still being built,
        422 when one failed or is unsupported).


        The answer lists `columnsAdded` and `columnsRemoved` (names ignore case,
        so a renamed header is one of each), `typeChanges` (a column whose
        dominant type changed), `nullPercentChanges` (empty cells moved by 5
        percentage points or more), the change in rows, columns, duplicate rows
        and quality score, and `schemaChanged`: true when a column was removed
        or retyped — the changes that break a reader of the data. New columns
        and shifts in blanks do not set it.
  /files/{id}:
    get:
      operationId: FilesController_get
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
      security: *ref_0
      tags: *ref_1
      summary: Get a file
      description: >-
        A file the caller cannot see is **404**, never 403, so its existence is
        not disclosed. The uploader and admins also get `grantedUserIds`, the
        colleagues a restricted file is shared with.
    delete:
      operationId: FilesController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDto'
      security: *ref_0
      tags: *ref_1
      summary: Delete a file
      description: >-
        Uploader or admin only. A soft delete: the file disappears from every
        listing and its stored object is removed, but the row and its history
        are kept and the upload still counts toward this period's quota.
    patch:
      operationId: FilesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
      security: *ref_0
      tags: *ref_1
      summary: Change who can see a file
      description: >-
        Uploader or admin only. `visibility` switches between `company` and
        `restricted`; `grantedUserIds` REPLACES the set of colleagues granted
        access (only for a restricted file — switching to `company` clears it).
        A colleague who can see the file but did not upload it gets 403; someone
        who cannot see it gets 404. The change takes effect on the next request
        and is audited.
  /files/{id}/download:
    get:
      operationId: FilesController_download
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownloadDto'
      security: *ref_0
      tags: *ref_1
      summary: Get a download link
      description: >-
        Checks the caller's access, then returns a signed link that expires in
        five minutes (`expiresAt`). The link needs no Authorization header; the
        browser saves it under the file's original name. With S3 storage it is a
        presigned S3 URL; with the development `local` driver it points at this
        API.
  /files/{id}/report:
    get:
      operationId: FilesController_report
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDto'
      security: *ref_0
      tags: *ref_1
      summary: Get a file's data-quality report
      description: >-
        Every upload queues a report, so this is never a 404 for a file you can
        see. `status` moves `queued` → `profiling` → `ready`. When `ready`,
        `metrics` holds the row and column counts and, per column: the share of
        empty cells, the inferred type (`integer`, `number`, `boolean`, `date`,
        `string`), whether the column mixes kinds of value and by how much, and
        numeric min / max / mean; plus duplicate and empty rows, ragged rows and
        header problems. Only the first 100,000 rows are profiled;
        `metrics.truncated` says when the file was longer.


        `narrative` is a plain-language summary and recommendations written by
        an AI model from those statistics alone — it is never shown a row or a
        cell value. It is `null` when no AI provider is configured, or it timed
        out or answered unusably; the metrics are unaffected either way.


        `unsupported` means the file is stored but not profiled (legacy `.xls`:
        save it as `.xlsx` or `.csv`). `failed` means it could not be read
        (corrupt, malformed, or too large to open safely) — `errorMessage` says
        why; a report can also show `failed` briefly while a temporary problem
        is retried automatically. Visible exactly when the file is: a file you
        cannot see is a 404.


        **Quality rules.** When the company has data-quality rules
        (`/quality-rules`), `ruleResults` holds one result per enabled rule —
        `passed`, `failed`, or `skipped` when the rule does not apply to this
        file (its column is not in it, or holds no numbers) — each with the rule
        as it was when the report was built, and `qualityScore` (0–100) is the
        share of applicable rules passed, an `error` rule counting double a
        `warning`. Both are `null` when the company had no rules, and the score
        is also `null` when none of them applied. Editing or deleting a rule
        later does not change an existing report: rebuild it (`POST
        /files/{id}/report/rebuild`) to check it against the rules as they are
        now.
  /files/{id}/report/rebuild:
    post:
      operationId: FilesController_rebuildReport
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportDto'
      security: *ref_0
      tags: *ref_1
      summary: Rebuild a file's report against today's rules
      description: >-
        Builds the report again from the same stored file, checking it against
        the company's quality rules as they are **now** — how an old file is
        re-checked after its rules changed. The uploader or an admin only
        (someone who can see the file but did not upload it gets 403; a file you
        cannot see is a 404). The report goes back to `queued` at once and the
        new result arrives as before (`file.status` over the realtime
        connection, or read the report again). Answers 409 while a build is
        already queued or running, so it cannot be requested twice. Needs the
        `files:write` scope for an API key.
  /files/{id}/preview:
    get:
      operationId: FilesController_preview
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreviewDto'
      security: *ref_0
      tags: *ref_1
      summary: Preview a file's first rows
      description: >-
        The first 50 rows of the file and the inferred type of each column, for
        rendering a table without downloading anything. Served from what
        profiling stored, so it costs no file read. Dates are ISO strings and
        long cells are cut to 200 characters. Returns 409 while the report is
        still being prepared, and 422 with the reason when no preview can be
        made (`unsupported` or `failed`). Visible exactly when the file is.
info:
  title: Gridline API
  description: >-
    Multi-tenant SaaS backend for spreadsheet upload, per-file permissions, and
    seat + usage billing.
  version: 0.1.0
  contact: {}
...
servers:
  - url: /api
  - url: http://localhost:4000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UploadFileBodyDoc:
      type: object
      properties:
        visibility:
          type: string
          enum: &ref_4
            - company
            - restricted
          default: company
        grantedUserIds:
          description: >-
            With `restricted`: the colleagues who may also see the file (the
            uploader and admins always can). Repeat the field, or send a JSON
            array. Empty means the uploader and admins only.
          type: array
          items:
            type: string
        file:
          type: string
          format: binary
          description: A CSV, XLS or XLSX file, up to 25 MB.
      required:
        - file
    FileDto:
      type: object
      properties:
        id:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
          description: From the file's bytes, not from what the client claimed.
        sizeBytes:
          type: number
        visibility:
          type: string
          enum: *ref_4
        datasetId:
          type: string
          format: uuid
          description: >-
            Shared by every version of one file. A first upload's is its own
            `id`.
        version:
          type: number
          description: >-
            Which version of the file this is: 1, 2, 3… Numbers are never
            reused.
        isLatest:
          type: boolean
          description: The newest version. A default list shows only these.
        uploaderId:
          type: string
        grantedUserIds:
          nullable: true
          description: >-
            Who a restricted file is shared with. Shown only to the uploader and
            admins, and only on single-file responses; null everywhere else.
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - originalName
        - mimeType
        - sizeBytes
        - visibility
        - datasetId
        - version
        - isLatest
        - uploaderId
        - grantedUserIds
        - createdAt
    UploadVersionBodyDoc:
      type: object
      properties:
        file:
          type: string
          format: binary
          description: A CSV, XLS or XLSX file, up to 25 MB.
      required:
        - file
    FileVersionPageDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FileDto'
        meta:
          $ref: '#/components/schemas/OffsetMetaDto'
      required:
        - data
        - meta
    ComparedFileDto:
      type: object
      properties:
        fileId:
          type: string
          format: uuid
        version:
          type: number
        originalName:
          type: string
      required:
        - fileId
        - version
        - originalName
...
    ComparisonDto:
      type: object
      properties:
        from:
          $ref: '#/components/schemas/ComparedFileDto'
        to:
          $ref: '#/components/schemas/ComparedFileDto'
        columnsAdded:
          description: >-
            In `to` but not in `from` (names ignore case; a renamed header is
            one removed and one added).
          type: array
          items:
            type: string
        columnsRemoved:
          description: In `from` but not in `to`.
          type: array
          items:
            type: string
        typeChanges:
          description: >-
            Columns whose dominant type changed. A column that is empty on
            either side has no type to compare.
          type: array
          items:
            $ref: '#/components/schemas/TypeChangeDto'
        nullPercentChanges:
          description: >-
            Columns whose share of empty cells moved by 5 percentage points or
            more.
          type: array
          items:
            $ref: '#/components/schemas/NullPercentChangeDto'
        rowCount:
          $ref: '#/components/schemas/CountChangeDto'
        columnCount:
          $ref: '#/components/schemas/CountChangeDto'
        duplicateRows:
          $ref: '#/components/schemas/CountChangeDto'
        qualityScore:
          description: The quality score; `delta` is null when either file was not scored.
          allOf:
            - $ref: '#/components/schemas/CountChangeDto'
        schemaChanged:
          type: boolean
          description: >-
            A column was removed or changed type — what breaks a reader of the
            data. New columns and shifts in blanks do not count.
      required:
        - from
        - to
        - columnsAdded
        - columnsRemoved
        - typeChanges
        - nullPercentChanges
        - rowCount
        - columnCount
        - duplicateRows
        - qualityScore
        - schemaChanged
    CursorMetaDto:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
          description: Pass as `cursor` to get the next page; null on the last page.
        hasMore:
          type: boolean
      required:
        - nextCursor
        - hasMore
    FilePageDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/FileDto'
        meta:
          $ref: '#/components/schemas/CursorMetaDto'
      required:
        - data
        - meta
    FileDownloadDto:
      type: object
      properties:
        url:
          type: string
          description: >-
            A short-lived link. Fetch it directly; it needs no Authorization
            header.
        expiresAt:
          type: string
          format: date-time
          description: The link stops working at this instant.
      required:
        - url
        - expiresAt
...
    ReportDto:
      type: object
      properties:
        fileId:
          type: string
        status:
          type: string
          enum:
            - queued
            - profiling
            - ready
            - unsupported
            - failed
        metrics:
          nullable: true
          description: Present once `status` is `ready`.
          type: object
          allOf:
            - $ref: '#/components/schemas/MetricsDto'
        narrative:
          nullable: true
          description: >-
            The plain-language summary. Null when no AI provider is configured
            or it produced nothing usable; the metrics are unaffected.
          type: object
          allOf:
            - $ref: '#/components/schemas/NarrativeDto'
        errorMessage:
          type: string
          nullable: true
          description: Why the report is `failed` or `unsupported`.
        profiledAt:
          type: string
          format: date-time
          nullable: true
        qualityScore:
          type: number
          nullable: true
          description: >-
            How the file did against the company’s quality rules, 0–100: the
            share of applicable rules it passed, an `error` rule counting double
            a `warning`. Null when no rule applied (or none is defined).
        ruleResults:
          nullable: true
          description: >-
            One result per rule the company had when this report was built, each
            carrying the rule as it was then (rules can be edited later). Null
            when the company had no rules. Rebuild the report to check against
            today’s rules.
          type: array
          items:
            $ref: '#/components/schemas/RuleResultDto'
      required:
        - fileId
        - status
        - metrics
        - narrative
        - errorMessage
        - profiledAt
        - qualityScore
        - ruleResults
...
    PreviewDto:
      type: object
      properties:
        columns:
          type: array
          items:
            $ref: '#/components/schemas/PreviewColumnDto'
        rows:
          type: array
          items:
            type: array
            items:
              nullable: true
              oneOf:
                - type: string
                - type: number
                - type: boolean
          description: >-
            The first rows, one array of cells per row in column order. Dates
            are ISO strings.
        totalRows:
          type: number
          description: Rows in the file (up to the profile budget).
        truncated:
          type: boolean
          description: True when `rows` is only the first part of the file.
      required:
        - columns
        - rows
        - totalRows
        - truncated
    UpdateFileDto:
      type: object
      properties:
        visibility:
          type: string
          enum: *ref_4
        grantedUserIds:
          description: >-
            REPLACES the set of colleagues granted access. Only meaningful for a
            `restricted` file; switching to `company` clears it.
          type: array
          items:
            type: string
    MessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: If that account exists, an email is on its way.
      required:
        - message
...